删除与编辑功能

+++ models/posts.js
@@ -1,6 +1,29 @@
 const marked = require('marked');
 const Post = require('../lib/mongo').Post;
 
+const CommentModel = require('./comments');
+
+//给post 添加留言数 commentsCount
+Post.plugin('addCommentsCount', {
+    afterFind: (posts)=>{
+        return Promise.all(posts.map((post)=>{
+            return CommentModel.getCommentsCount(post._id).then((commentsCount)=>{
+                post.commentsCount = commentsCount;
+                return post;
+            });
+        }));
+    },
+    afterFindOne: (post)=>{
+        if(post){
+            return CommentModel.getCommentsCount(post._id).then((count){
+               post.commentsCount = count;
+               return post;
+            });
+        }
+        return post;
+    }
+})
+
 // 将 post 的 content 从 markdown 转换成 html
 Post.plugin('contentToHtml', {
     afterFind: (posts)=>{
@@ -63,8 +86,14 @@ module.exports = {
         return Post.update({_id: postId}, {$set: data}).exec();
     },
     //通过文章id删除一篇文章
-    delPostById: (postId)=>{
-        return Post.deleteOne({_id: postId}).exec();
+    delPostById: (postId, author)=>{
+        return Post.deleteOne({author: author, _id: postId}).exec()
+            .then((res)=>{
+                //文章删除后，再删除该文章下的所有留言
+                if(res.result.ok && res.result.n > 0){
+                    return CommentModel.delCommentsByPostId(postId);
+                }
+            });
     }
 
 }
